Refactor: IdCard form 관련 type 수정
