+++ src/modules/IdCardCreation/IdCardCreation.type.ts
@@ -1,0 +1,1 @@
+export type CreationSteps = 'BOARDING' | 'PROFILE' | 'KEYWORD' | 'KEYWORD_CONTENT' | 'COMPLETE';
