+++ src/modules/IdCardCreation/Step/KeywordContentStep.client.tsx
@@ -2,7 +2,7 @@
 import { useFormContext } from 'react-hook-form';
 
 import { KeywordContentEditCard } from '~/modules/IdCardCreation/Card';
-import { CreateKeywordModel } from '~/types/idCard';
+import { FormKeywordModel } from '~/types/idCard';
 
 const title = '나를 소개하는 키워드의\n 설명을 적어주세요!';
 
@@ -14,7 +14,7 @@ export const KeywordContentStep = () => {
     <div>
       <h1 className="text-h1">{title}</h1>
       <div className="mt-[26px] flex flex-col gap-[22px]">
-        {keywords.map((keyword: CreateKeywordModel, index: number) => {
+        {keywords.map((keyword: FormKeywordModel, index: number) => {
           return <KeywordContentEditCard key={index} keyword={keyword} index={index} />;
         })}
       </div>
