+++ src/modules/IdCardEditor/IdCardEditor.type.ts
@@ -1,0 +1,1 @@
+export type EditorSteps = 'PROFILE' | 'KEYWORD' | 'KEYWORD_CONTENT';
